user_group_settings_ui: Merge into user_group_edit.
