+++ tools/test-js-with-node
@@ -257,7 +257,6 @@ EXEMPT_FILES = make_set(
         "web/src/user_group_popover.js",
         "web/src/user_group_ui_updates.js",
         "web/src/user_groups.ts",
-        "web/src/user_groups_settings_ui.js",
         "web/src/user_profile.js",
         "web/src/user_settings.ts",
         "web/src/user_sort.js",
